requirements.txt
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
attrs==21.4.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
chart-studio==1.1.0
click==8.1.3
colorama==0.4.5
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
greenlet==1.1.2
html5lib==1.1
idna==3.3
imbalanced-learn==0.9.1
importlib-metadata==6.0.0
install==1.3.5
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jupyterlab_server==2.19.0
kiwisolver==1.4.3
lxml==4.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
nbclassic==0.5.1
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
notebook_shim==0.2.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.6.2
plotly==5.8.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
pydot==1.4.2
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==25.0.0
regex==2022.6.2
requests==2.28.0
retrying==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
stack-data==0.3.0
streamlit==1.19.0
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.9
validators==0.20.0
watchdog==2.3.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
widgetsnbextension==3.6.0
zipp==3.12.0
